+++ Apigateway/models/Deployment.cs
@@ -91,6 +91,10 @@ namespace Oci.ApigatewayService.Models
         [JsonProperty(PropertyName = "endpoint")]
         public string Endpoint { get; set; }
         
+        /// <remarks>
+        /// Required
+        /// </remarks>
+        [Required(ErrorMessage = "Specification is required.")]
         [JsonProperty(PropertyName = "specification")]
         public ApiSpecification Specification { get; set; }