+++ Oda/OdaWaiters.cs
@@ -60,6 +60,33 @@ namespace Oci.OdaService
         /// <param name="request">Request to send.</param>
         /// <param name="targetStates">Desired resource states. If multiple states are provided then the waiter will return once the resource reaches any of the provided states</param>
         /// <returns>a new Oci.common.Waiter instance</returns>
+        public Waiter<GetOdaInstanceAttachmentRequest, GetOdaInstanceAttachmentResponse> ForOdaInstanceAttachment(GetOdaInstanceAttachmentRequest request, params OdaInstanceAttachment.LifecycleStateEnum[] targetStates)
+        {
+            return this.ForOdaInstanceAttachment(request, WaiterConfiguration.DefaultWaiterConfiguration, targetStates);
+        }
+
+        /// <summary>
+        /// Creates a waiter using the provided configuration.
+        /// </summary>
+        /// <param name="request">Request to send.</param>
+        /// <param name="config">Wait Configuration</param>
+        /// <param name="targetStates">Desired resource states. If multiple states are provided then the waiter will return once the resource reaches any of the provided states</param>
+        /// <returns>a new Oci.common.Waiter instance</returns>
+        public Waiter<GetOdaInstanceAttachmentRequest, GetOdaInstanceAttachmentResponse> ForOdaInstanceAttachment(GetOdaInstanceAttachmentRequest request, WaiterConfiguration config, params OdaInstanceAttachment.LifecycleStateEnum[] targetStates)
+        {
+            var agent = new WaiterAgent<GetOdaInstanceAttachmentRequest, GetOdaInstanceAttachmentResponse>(
+                request,
+                request => client.GetOdaInstanceAttachment(request),
+                response => targetStates.Contains(response.OdaInstanceAttachment.LifecycleState.Value)
+            );
+            return new Waiter<GetOdaInstanceAttachmentRequest, GetOdaInstanceAttachmentResponse>(config, agent);
+        }
+        /// <summary>
+        /// Creates a waiter using default wait configuration.
+        /// </summary>
+        /// <param name="request">Request to send.</param>
+        /// <param name="targetStates">Desired resource states. If multiple states are provided then the waiter will return once the resource reaches any of the provided states</param>
+        /// <returns>a new Oci.common.Waiter instance</returns>
         public Waiter<GetWorkRequestRequest, GetWorkRequestResponse> ForWorkRequest(GetWorkRequestRequest request, params WorkRequest.StatusEnum[] targetStates)
         {
             return this.ForWorkRequest(request, WaiterConfiguration.DefaultWaiterConfiguration, targetStates);