+++ Oda/models/CreateOdaInstanceDetails.cs
@@ -66,8 +66,7 @@ namespace Oci.OdaService.Models
         public System.Nullable<ShapeNameEnum> ShapeName { get; set; }
         
         /// <value>
-        /// Simple key-value pair that is applied without any predefined name, type or scope. Exists for
-        /// cross-compatibility only.
+        /// Simple key-value pair that is applied without any predefined name, type, or scope.
         /// Example: {&quot;bar-key&quot;: &quot;value&quot;}
         /// </value>
         [JsonProperty(PropertyName = "freeformTags")]
@@ -79,6 +78,18 @@ namespace Oci.OdaService.Models
         /// </value>
         [JsonProperty(PropertyName = "definedTags")]
         public System.Collections.Generic.Dictionary<string, System.Collections.Generic.Dictionary<string, System.Object>> DefinedTags { get; set; }
+        
+        /// <value>
+        /// Should this Digital Assistant instance use role-based authorization via an identity domain (true) or use the default policy-based authorization via IAM policies (false)
+        /// </value>
+        [JsonProperty(PropertyName = "isRoleBasedAccess")]
+        public System.Nullable<bool> IsRoleBasedAccess { get; set; }
+        
+        /// <value>
+        /// If isRoleBasedAccess is set to true, this property specifies the identity domain that is to be used to implement this type of authorzation. Digital Assistant will create an Identity Application instance and Application Roles within this identity domain. The caller may then perform and user roll mappings they like to grant access to users within the identity domain.
+        /// </value>
+        [JsonProperty(PropertyName = "identityDomain")]
+        public string IdentityDomain { get; set; }
         
     }
 }