+++ Oda/models/UpdateOdaInstanceDetails.cs
@@ -34,8 +34,7 @@ namespace Oci.OdaService.Models
         public string Description { get; set; }
         
         /// <value>
-        /// Simple key-value pair that is applied without any predefined name, type or scope. Exists for
-        /// cross-compatibility only.
+        /// Simple key-value pair that is applied without any predefined name, type, or scope.
         /// Example: {&quot;bar-key&quot;: &quot;value&quot;}
         /// </value>
         [JsonProperty(PropertyName = "freeformTags")]