+++ Oda/responses/CreateOdaInstanceResponse.cs
@@ -31,7 +31,7 @@ namespace Oci.OdaService.Responses
 
         /// <value>
         /// Unique Oracle-assigned identifier for the asynchronous request. You can use this to query status
-        /// of the asynchronous operation.
+        /// of the operation.
         /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "opc-work-request-id")]
@@ -39,8 +39,8 @@ namespace Oci.OdaService.Responses
 
 
         /// <value>
-        /// Unique Oracle-assigned identifier for the request. If you contact Oracle
-        /// about this request, provide this request ID.
+        /// Unique Oracle-assigned identifier for the request. If you need to contact
+        /// Oracle about a particular request, please provide the request ID.
         /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "opc-request-id")]