+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a [Changelog](http://keepachangelog.com/).
 
+## 38.2.0 - 2022-05-24
+### Added
+- Support for the License Manager service
+- Support for usage plans in the API Gateway service
+- Support for packaged skill and instance metadata management, role-based access options on instance creation, and assigned ownership in the Digital Assistant service
+- Support for compute capacity reservations in the VMWare Solution service
+- Support for Oracle Linux 8 application streams in the OS Management service
+ 
+### Fixed
+- Service Error message was fixed to log correct Service name
+
 ## 38.1.0 - 2022-05-17
 ### Added
 - Support for information requests in the Operator Access Control service